Refactor ConstructStructAnnotation to handle union
ConstructStructAnnotation can now handle most union cases.
This change stops the function from iterating through every field of
the RD when calculating CBufferSize/Offset and building
fieldAnnotations for unions.
Now when handling unions it finds the appropriate field matching the
structType in the annotation, and only builds a fieldAnnotation for
that field.

Currently bitfields are not handled correctly in unions in
ConstructStructAnnotation.